Modularized code a bit. Fixed some import bugs. Added `ls`.
